requirements.txt
appnope==0.1.0
bleach==2.1.3
bokeh==0.12.14
certifi==2018.1.18
click==6.7
cycler==0.10.0
decorator==4.2.1
entrypoints==0.2.3
Flask==0.12.2
Flask-HTTPAuth==3.2.3
html5lib==1.0.1
ipykernel==4.8.2
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.2
itsdangerous==0.24
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.2
jupyter-console==5.2.0
jupyter-core==4.4.0
kiwisolver==1.0.1
MarkupSafe==1.0
matplotlib==2.2.0
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.4.0
numpy==1.14.2
packaging==17.1
pandas==0.22.0
pandocfilters==1.4.2
parso==0.1.1
patsy==0.5.0
pexpect==4.4.0
pickleshare==0.7.4
pipenv==11.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.7.0
python-dotenv==0.8.2
pytz==2018.3
PyYAML==3.12
pyzmq==17.0.0
qtconsole==4.3.1
scipy==1.0.0
seaborn==0.8.1
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
statsmodels==0.8.0
terminado==0.8.1
testpath==0.3.1
tornado==5.0
traitlets==4.3.2
virtualenv==15.1.0
virtualenv-clone==0.3.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.1.4
